+++ apps/bbsync/tests/test_integration.py
@@ -513,6 +513,7 @@ class TestBBSyncIntegration:
         bugzilla_token,
         enable_bz_sync,
         enable_jira_tracker_sync,
+        jira_email,
         jira_token,
         test_api_v2_uri,
     ):
@@ -673,6 +674,7 @@ class TestBBSyncIntegration:
             flaw_data,
             format="json",
             HTTP_BUGZILLA_API_KEY=bugzilla_token,
+            HTTP_JIRA_API_EMAIL=jira_email,
             HTTP_JIRA_API_KEY=jira_token,
         )
         assert response.status_code == 200
@@ -1066,10 +1068,8 @@ class TestFlawDraftBBSyncIntegration:
         assert flaw.title == f"From {source} collector"
 
         # check values related to taskman
-        assert flaw.group_key == ""
         assert flaw.owner == ""
         assert flaw.task_key
-        assert flaw.team_id == ""
         assert flaw.workflow_name == "DEFAULT"
         assert flaw.workflow_state == "NEW"