+++ apps/workflows/workflow.py
@@ -13,6 +13,8 @@ from typing import Optional
 import yaml
 from django.db import models
 
+from osidb.helpers import deprecate_field
+
 from .constants import WORKFLOW_DIR
 from .exceptions import (
     InitialStateException,
@@ -189,10 +191,10 @@ class WorkflowModel(models.Model):
         default=WorkflowState.NOVALUE,
     )
     owner = models.CharField(max_length=60, blank=True)
-    group_key = models.CharField(max_length=60, blank=True)
+    group_key = deprecate_field(models.CharField(max_length=60, blank=True))
     task_key = models.CharField(max_length=60, blank=True)
     task_updated_dt = models.DateTimeField(null=True, blank=True)
-    team_id = models.CharField(max_length=8, blank=True)
+    team_id = deprecate_field(models.CharField(max_length=8, blank=True))
 
     class Meta:
         abstract = True
@@ -316,7 +318,7 @@ class WorkflowModel(models.Model):
         """
         WorkflowFramework().validate_classification(self, target_workflow, target_state)
 
-    def promote(self, save=True, jira_token=None, **kwargs):
+    def promote(self, save=True, jira_token=None, jira_email=None, **kwargs):
         """
         this is the cannonical way of changing classification
 
@@ -332,11 +334,12 @@ class WorkflowModel(models.Model):
         if save:
             self.save(
                 jira_token=jira_token,
+                jira_email=jira_email,
                 raise_validation_error=False,
                 **kwargs,
             )
 
-    def revert(self, save=True, jira_token=None, **kwargs) -> None:
+    def revert(self, save=True, jira_token=None, jira_email=None, **kwargs) -> None:
         """
         This is the canonical way of reverting to a previous valid state.
         """
@@ -348,11 +351,12 @@ class WorkflowModel(models.Model):
         if save:
             self.save(
                 jira_token=jira_token,
+                jira_email=jira_email,
                 raise_validation_error=False,
                 **kwargs,
             )
 
-    def reset(self, save=True, jira_token=None, **kwargs) -> None:
+    def reset(self, save=True, jira_token=None, jira_email=None, **kwargs) -> None:
         """
         This is the canonical way of resetting to the default workflow.
         """
@@ -360,9 +364,14 @@ class WorkflowModel(models.Model):
 
         self.classification = (default_workflow, WorkflowModel.WorkflowState.NEW)
         if save:
-            self.save(jira_token=jira_token, raise_validation_error=False, **kwargs)
+            self.save(
+                jira_token=jira_token,
+                jira_email=jira_email,
+                raise_validation_error=False,
+                **kwargs,
+            )
 
-    def reject(self, save=True, jira_token=None, **kwargs):
+    def reject(self, save=True, jira_token=None, jira_email=None, **kwargs):
         """
         this is the cannonical way of rejecting a flaw / task
 
@@ -379,6 +388,7 @@ class WorkflowModel(models.Model):
         if save:
             self.save(
                 jira_token=jira_token,
+                jira_email=jira_email,
                 raise_validation_error=False,
                 **kwargs,
             )